+++ src/pages/additional/index.js
@@ -1,5 +1,6 @@
 import './games/game';
 import './writer';
+import './gallery/galleryPosters';
 import './gallery/gallery';
 import './map';
 import './tabs';