+++ src/services/api/envService.js
@@ -12,4 +12,8 @@ export default class EnvService {
     static getFlickrAPIKey() {
         return EncryptService.decodeBase64Str(process.env.FLICKR_API_KEY);
     }
+    
+    static getTheMovieDBAPIKey() {
+        return EncryptService.decodeBase64Str(process.env.THEMOVIEDB_API_KEY);
+    }
 }