+++ Changelog.md
@@ -2,6 +2,7 @@
 v 3.01
   Added GameService to the Game, Safe Game pages
   Improved TextEditor, Games viewer design, left, right sides
+  Created Gallery TheMoviedb Posters page
 
 v 3.00
   Created numeraction for FAQ component